changed id type from Long to BigInteger

+++ src/main/java/com/bloomtech/testapi/repositories/EmployeeRepository.java
@@ -4,8 +4,10 @@ import com.bloomtech.testapi.entities.Employee;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigInteger;
+
 @Repository
-public interface EmployeeRepository extends JpaRepository<Employee, Long> {
+public interface EmployeeRepository extends JpaRepository<Employee, BigInteger> {
     public Employee findEmployeeByFirstName(String firstName);
 
 }
